Some conversion to a module-based system.

+++ extract_date2id.py
@@ -11,17 +11,24 @@
 #
 
 import os, json
-lsdir=[]
-for r,ds,fs in os.walk(r"..\..\egs"):
-    if ".git" not in r:
-        lsdir.extend(fs)
-pl=len("ST-2010-04-23-")
-def getdate(fn):
-    return fn[3:][:10],int(fn[pl:][:5])
 
-stmap=dict(map(getdate,filter(lambda i:i.startswith("ST-") and not i.lower().count("-original") and not i.lower().count("-remastered-sb"),lsdir)))
-npmap=dict(map(getdate,filter(lambda i:i.startswith("NP-") and not i.lower().count("-original") and not i.lower().count("-colour-official"),lsdir)))
-sbmap=dict(map(getdate,filter(lambda i:i.startswith("SB-") and not i.lower().count("-original") and not i.lower().count("-xxxxx"),lsdir)))
+def extract_date2id():
+    print (">>> extract_date2id")
+    if not os.path.exists(r"..\..\egs\0003 - (Sketchbook and Extras)\0001 - Filler Sketchbook"):
+        print ("Not running on HarJIT's machine, not regenerating Date2Id.txt")
+        return
+    lsdir=[]
+    for r,ds,fs in os.walk(r"..\..\egs"):
+        if ".git" not in r:
+            lsdir.extend(fs)
+    pl=len("ST-2010-04-23-")
+    def getdate(fn):
+        return fn[3:][:10],int(fn[pl:][:5])
+    stmap=dict(map(getdate,filter(lambda i:i.startswith("ST-") and not i.lower().count("-original") and not i.lower().count("-remastered-sb"),lsdir)))
+    npmap=dict(map(getdate,filter(lambda i:i.startswith("NP-") and not i.lower().count("-original") and not i.lower().count("-colour-official"),lsdir)))
+    sbmap=dict(map(getdate,filter(lambda i:i.startswith("SB-") and not i.lower().count("-original") and not i.lower().count("-xxxxx"),lsdir)))
+    if stmap and npmap and sbmap:
+        open("Date2Id.txt","w").write(json.dumps({"story":stmap,"sketch":sbmap,"np":npmap}))
 
-if stmap and npmap and sbmap:
-    open("Date2Id.txt","w").write(json.dumps({"story":stmap,"sketch":sbmap,"np":npmap}))
+if __name__=="__main__":
+    extract_date2id()
